final сalculator

+++ kalkulator/Program.cs
@@ -4,21 +4,22 @@ namespace kalkulator
 {
     class Program
     {
+
         static void Main(string[] args)
         {
             Console.WriteLine("Enter first value:");
-            int number1 = int.Parse(Console.ReadLine());
+            double number1 = double.Parse(Console.ReadLine());
 
             Console.WriteLine("Enter second value:");
-            int number2 = int.Parse(Console.ReadLine());
+            double number2 = double.Parse(Console.ReadLine());
 
             Console.WriteLine("Enter symbol:");
             char symbol = char.Parse(Console.ReadLine());
 
-            int plus = number1 + number2;
-            int minus = number1 - number2;
-            int multiplication = number1 * number2;
-            int division = number1 / number2;
+            double plus = number1 + number2;
+            double minus = number1 - number2;
+            double multiplication = number1 * number2;
+            double division = number1 / number2;
 
             if (symbol == '+')
             {
